migrate unit tests to new transform/glue_test module
